Restructure the cron script and fix bugs.
Do not inherit from the Config class, it only made things more complicated.
Remove the sys.path.append that was made necessary by using partial paths where full
paths are expected.
Exit with 1 (to make commit fail properly) if VyOSError is raised.
Replace the OO representation of cron jobs with simple hashes: the format is not expected
to grow any more complex and encapsulation will hardly be necessary (though if one can think of
a good OO design for this, it's welcome).
Do not mix config reading and validation to enable changing the syntax without changing any of the
code but get_config()