bugfix: DNS domain forward server syntax
When generating a configuration for DNS domain forward to a local server
with 'set service dns forwarding domain foo.com server 1.1.1.1' this resulted in

  # domain ... server ...
  forward-zones=bar.com = 2.1.1.1;2.1.1.2
  forward-zones=foo.com = 1.1.1.1;1.1.1.2

On PowerDNS recursor restart the last line won and it totally forgot about
'bar.com'. This could be seen from the logfiles that only one domain was loaded:

  Redirecting queries for zone 'foo.com.' to: 1.1.1.1:53, 1.1.1.2:53

The manual at https://doc.powerdns.com/3/recursor/settings/#forward-zones shows
that all domains have to be configured on one 'forward-zones=' line. In the above
example this has to result in:

  forward-zones=bar.com=2.1.1.1;2.1.1.2, foo.com=1.1.1.1;1.1.1.2

A subsequent check within the logfiles reveal that it's now working:

  Redirecting queries for zone 'bar.com.' to: 2.2.2.2:53, 2.2.2.1:53
  Redirecting queries for zone 'foo.com.' to: 1.1.1.1:53, 1.1.1.2:53