T998: "service dns dynamic" does now honor the "use-web" statement
This bug was present since the old Vyatta days as the use-web statement
was only put into action when also "use-web skip" was defined.

The service https://ipinfo.io/ip does not place any crap in front of the
IP address so the skip statement was not used and made no sense.