powerdns: T1524: support setting allow-from network
Netmasks (both IPv4 and IPv6) that are allowed to use the server. The default
allows access only from RFC 1918 private IP addresses. Due to the aggressive
nature of the internet these days, it is highly recommended to not open up the
recursor for the entire internet. Questions from IP addresses not listed here
are ignored and do not get an answer.

https://docs.powerdns.com/recursor/settings.html#allow-from

Imagine an ISP network with non RFC1918 IP adresses - they can't make
use of PowerDNS recursor.

As of now VyOS hat allow-from set to 0.0.0.0/0 and ::/0 which created an open
resolver. If there is no allow-from statement a config-migrator will add
the appropriate nodes to the configuration, resulting in:

 service {
     dns {
         forwarding {
             allow-from 0.0.0.0/0
             allow-from ::/0
             cache-size 0
             ignore-hosts-file
             listen-address 192.0.2.1
         }
     }
 }

(cherry picked from commit dc0f641956d002fa8588ef8d1213791cf36e92f2)