bonding: T1614: T532: new commit validators
As in the past during the priority race of the bash script invalid configuration
could appear in the CLI and are de-synced from the kernle state, e.g. some
bonding modes do not support arp_interval.

This is no longer allowed and added to the migration script so that the config
again represents the truth.