bonding: T1660: bugfix for triggered OS permission denied exception
Some bond modes do not support arp monitor interval and thus internally eset it
to 0 which means disabled. If you then write to the sysfs file in question an
OS exception (permission denied) is triggered.

arp_mon_intvl is initialized as 0 which means disabled so we only write it
when it is really requested by the user. There is a validator ensuring that
it can only be set in the bond modes which support it.