Python/ifconfig: T1712: wait when changing interface state
With some interfaces, for example bond vif, it take some time for the state
change to really happen. Because of this later code, like starting DHCP client,
might not work as expected as get_state() reports the old (real) state.

Now when changing state of an interface we are (busy-)waiting up to 12.5
seconds before we inform the user that the interface could not be brought up.
This should be more then enough time for any interface to start except when
there is really no cable attached.