Python/ifconfig: T1712: always start DHCP when configured
DHCP was only started when the interface operstate was set to up but this is
wrong. An interface can be configured as DHCP interface and DHCP client must
be launched even when the physical interface becomes available n-minutes
later. DHCP client then can ask for an IP assignemnt by DHCP.

Tested by starting DHCP client on a not connected ethernet interface. Address
was assigned later on after link became ready.