ifconfig: T1793: add delta check on set_flow_control()
The flow control settings should only be changed when they need to. If flow
control is altered, the kernel will disable and re-enable the interface. This
will not only let the switchport flap but it will also reset e.g. BGP
sessions.

In addition - this also reduces the config commit time.