ifconfig: T1793: add delta check on set_speed_duplex()
The speend and duplex settings should only be changed when they need to.
Always configuring this setting will make the kernel disable and re-enable the
physical interface. This will not only let the switchport flap but it will also
reset e.g. BGP sessions.

This is the first part of this fix for speed/duplex auto settings.

In addition - this also reduces the config commit time.