vyos-hostsd: T1885: Fixed crash if stdout PIPE connection lost
In case if stdout PIPE were lost (for example, after systemd-journald restart), using print() will raise the exception.
This patch replaces print() to logging function, which is not sensitive to this problem. This makes vyos-hostd resistant to stdout disconnection, but of course, all logs after stdout disconnection will be lost.