VRRP: T1588: Added support for dynamic interfaces in VRRP
It is possible to add a nonexistent interface to the VRRP configuration, but this will lead to keepalived error and stop after configuration reload.
With this new parameter, keepalived will consider nonexistent interfaces as FAULT instead crash and start VRRP on them, when they will appear on the system, which is a feature and much more correct behavior.