dynamic DNS: T1953: Relaxed service name check
Internally, we can accept more than one server of each type for sending dynamic DNS updates, but due to a strong check in CLI, it is not possible to add more than one server with the same protocol (except "custom", but it allows to add only one more server). The patch relaxing this limitation by allowing adding as many servers with the same protocol, as needed.