vyos.validate: remove interface suffix when running is_ipv6_link_local()
Commit 17dd50751d ("bfd: T1949: fix verification logic for IPv6 BFD peers")
added a mechanism to probe if an IPv6 address is link-local or not. Sometimes
an interface suffix is appended by %interface to a link-local address, as we
need to know an interface for which this address is (hence the name) link-local.

Remove any interface identifier before checking if the address is local or not.