T2034: Disallow removal of default loopback addresses
The removal of interfaces loopback lo results in all address being
removed from the loopback interface. (also not cli controlled addresses)
In this process 127.0.0.1/8 and ::1/128 are also removed witch results
in error for services that are dependent on these adresses, this includes
eg. snmp and ssh

Removal of these addresses needs to be disallowed by the config backend
and removal of the whole config block interfaces loopback lo needs to
result in removal of all non-default addresses only.