interfaces: wireguard: add initial test
WireGuard interface can not be easily inherit the BasicInterfaceTest class
as every operation on WireGuard interfaces require IP addresses assigned
to the interface - I do not now if this is really a requirement or just
a to restrict validate() on the WireGuard interface.

Nevertheless this helps to see if changes on e.g. ifconfig.py breaks this
interface.