bridge: l2tpv3: T1823: readjust overall priorities
Every interface on the system can be a member of a bridge - thus the bridge
interface must be the one interface which has the highest priority
compared to all other interfaces - incl. l2tpv3.

With this change the system boots up fine with also l2tpv3 interfaces
participating as bridge members. This change was needed as a l2tpv3 interface
requires proper configured routing oin the system, else adding the
interface will fail miserably:

  FileNotFoundError: [Errno 2] ip l2tp add tunnel tunnel_id 200
     peer_tunnel_id 100 udp_sport 4000 udp_dport 3000 encap udp
     local 172.18.201.10 remote 172.18.203.10
  returned: RTNETLINK answers: Network is unreachable