wwan: migrate alias and vrf script from pre-up to up
As the pre-up script is only run once when the interface is brought up but
not when the interface maybe only comes up 20 seconds later due to the remote
site some actions can not be performed as placing the interface e.g. into a VRF
instance.

TODO: also move the backup default route parts to this new script.