vrf: T31: name of isntance is not allowed to mimic an interface name
Every VRF that's created is not allowed to be named like any interface that
can be active on the system. This includes eth, lan, br, dum, lo ....

In theoriy this would work but as soon as such a regular interface is created
things will go sideways rather quick thus we limit the namespace which can
be used to create a VRF.

Appending an interface name is still possible like coolvrf-eth0.