ifconfig: T2190: option to prevent Interface creation
a new option was added to the Interface class "create".

By default the value is set to True, and when an instance of the
class is created and the underlying interface does not exists, the
class will create it.

If the option "create" is set to False, the interface will not be
created and instead the class  will raise an error when it is
instantiated.