tunnel: T2028: register the _Tunnel class
All interface classes register themselves so that we can figure
out from an interface name what class is responsible for its
creation. In the case of Tunnel, it is not the case as multiple
different interfaces are sharing the same "tun" prefix.

Still many operations are the same and it should be done.