pppoe: T2248: connect-on-demand and vrf are mutually exclusive
As PPP can be used to establish a connection on-demand it manages the Kernel
default route. This can not be used when using VRFs which are managed by
the ip-up.d and ip-down.d scripts - thus those options are now mutially
exclusive.

The best fix would be adding support for VRFs into PPP.