configdict: T2427: do not remove all addresses when disabling interface
Commit 3fdf0093a introduced code that removed all addresses from an
interface when that interface is disabled. This is wrong, as other
configured services may be listening on these addresses and may fail to
start if their configured address isn't present.

It also caused a commit error when applying dhcp-server configuration:

    DHCP server configuration error!
    None of configured DHCP subnets does not have appropriate
    primary IP address on any broadcast interface.

This commit reverts it to prior behavior, which was to just put the
interface admin down and leave all addresses configured, other than the
IPv6 'fe80::EUI-64/64' link-local, which it deletes, as the interface
may not have a MAC if it's put down.