T2267: Versioning: Update version tag from GIT repo
This commit will update the version field of the generated deb package from the git repo version tag.
the tag needs to be in the format "vyos/<version>" eg. "vyos/1.2.5" 1.2.5 is then used as the version
The version field will be one of the following syntaxes:
on a commit:
    - <version>-<commits from tag>-g<commit id>
    eg. 1.2.5-4-g23232343

on a commit with unstaged changes:
    - <version>-<commits from tag>-g<commit id>+dirty
    eg. 1.2.5-4-g23232343+dirty

This will make it clear what state the repo was in when the package was generated.

It is possible to remove the number and commit id on the tagged commit, but i've decided to use the same
format for all commits tagged or not. as for now the tagged commit will be sufixed with -0-g<commit id>

if no valid tag is recieved from `git describe`, the script uses a 0.0 version number.

The changelog is also updated to reflext that we dont use it and refers to the Git Changelog and vyos release-notes