ifconfig/dhcp: T2449: remove accept_ra logic as it was wrong
Currently accept_ra was set to 0 if 'address dhcpv6' was set on an
interface. This is wrong, as without RA, the system will get no routes
to the DHCPv6-obtained prefix.

Since the logic for accept_ra was moved to the interface scripts,
it can be removed from the dhclient code.