http-api: T2494: remove "barrier" PreExec
The PreExec is making sure that the vyos-config-status file exists
and blocks until it does.

This file is created on boot completion and I can see no reason
why the http service has to wait for the end of boot to start.

Any barrier to start should be done with systemd itself.