udev: T2490: add persistent USB device files
During testing it was discovered that on 5 out of 10 reboots the USB
enumeration/mapping from physical port to /dev/ttyUSB is different. The root
cause is that it's a FIFO so first found/loaded driver module will be assigned
ttyUSB0.

This mixed up the serial interfaces of my FTDI chips and my connected Sierra
Wireless MC7710 card which was no longer functioning as it now was mapped to
a different USB interface.

The solution is a udev rule which persistently maps the USB-tree-device to a
device file in /dev. Wait? isn't this what /dev/serial/by-{id,path} is for?
Correct, it does the very same thing but the problem is as follows:

* by-path uses device file names which also incorporate the parent bus system,
  this results in "pci-0000:00:10.0-usb-0:2.4:1.0-port0"
* by-id will overwrite the assigned device symlink if a new USB device with the
  same name appears. This happens to some FTDI devices with no serial number
  programmed so the device added last wins and will be the only one in
  the by-id folder - cruel world!

This commit adds a new directory /dev/serial/by-bus which holds the following
device files (as example):

  $ ls -1 /dev/serial/by-bus/
  usb0b1.3p1.0
  usb0b1.3p1.2
  usb0b1.3p1.3
  usb0b2.4p1.0
  usb0b2.4p1.1
  usb0b2.4p1.2
  usb0b2.4p1.3