T2638: FRR: New framework for configuring FRR
This commit adds the python module vyos.frr that support reading and modifying the frr cofiguration.

Functions
get_configuration() :
    Collects the configuration from FRR and returns it as a string

reload_configuration() :
    uses frr-reload.py to activate a new configuration.
    The configuration applied will overwrite the current configuration.

mark_configuration() :
    Does syntax check/validation and add "end" tags to the configuration
    The marked configuration will be returned as a string
    A syntax fault will render an Exception

execute() :
    Execute a command inside vtysh

configure() :
    Executes a command in vtysh config mode

replace_section() / _replace_section() :
    Adds the ability to replace a section of frr code

remove_section() :
    Removes the specified configuration block from the config

For now this supports replacing complete config sub-blocks of configuration and selecting the daemon
to replace inside. This should work for most daemons, but static routing will still be an issue because
this is not a separate sub-config mode