ethernet: support testing via SSH
Instead of running the smoketest on all available interfaces it is now possible
to specify a whitespace separated list as environment variable where the test
is performed. This is especially usefull when developing/testing via SSH and
to limit the test to an interface you are not connected to.

Test only with interface eth1:
$ TEST_ETH='eth1' /usr/libexec/vyos/tests/smoke/cli/test_interfaces_ethernet.py