mpls-conf: T915: Add TTL security for static LDP neighbors
The commit has to do with the addition of TTL security for LDP neighbors. The code was 90% done by Viascheslav. I modified it a little bit to get it to properly work.

We added more parameters to the neighbors dynamic loop. Once this is merged then we should be able to add more for the dynamic neighbor statements.

The way that this works is to have either TTL disabled, or to add the amount of hops accepted for the neighbor.