ifconfig: T1405: ensure MAC address is configured first
The MAC address is changed after we have set an IP address on the interface or
started dhclient. This will cause some users to receive the wrong IP address
on device startup.

Change to order of how parameters are set in the system. The interface MAC
address is now configured first.