mpls-conf: T915: Refactored FRR LDP template, MPLS handler, added MPLS global features
So this is a big update.

The first thing that was done was a refactor to the FRR LDP template, MPLS handler, and XML conf tree MPLS global additions.

The refactors should work and I did test them in my lab. It seems that everything does work as needed so far in my testing.

There is something here that is considered configuration breaking from the old setup though. In the old setup the MPLS interface operation (as in the interfaces accepting MPLS labels and processing them) was tied with LDP. What this means is that MPLS processing was enabled at the same time as LDP interfaces were configured. We do not want this behavior for the future as there's other MPLS underlay technologies like SR and RSVP. If someone wants to enable SR or RSVP without enabling LDP then they now can. Before, they couldn't. The other additions are global changes to MPLS TTL propagation and MPLS max TTL enforcement. They have now been added.

Lastly, there is an frr-reload bug that Runar Borge found with this. We have found that when totally deleting LDP that there has to be 3 commits done. This is because frr-reload doesn't properly do what it needs to do in 1 operation so we had to do 3. This will only affect people that are doing an entire LDP clear using "delete protocols mpls ldp." Otherwise it isn't seen.

Anyway, this refactor now works with the FRR daemon directly for all changes. This also makes it much easier for adding stuff in the future.

Thank you